directory and file name refactoring
