+++ src/screens/Member/MemberProfile.module.css
@@ -16,25 +16,23 @@
   z-index: 1;
 }
 
-.MemberProfileButtons {
+.MemberProfileAdminButton {
   position: absolute;
   top: 2%;
-  right: 2%;
-  display: flex;
-  align-items: center;
-  column-gap: 15px;
-}
-
-.MemberProfileAdminButton {
-  padding: 7px 10px;
+  left: 1%;
+  padding: 7px 8px;
   background-color: #2d285c;
   color: white;
   border-radius: 10px;
   font-family: "Poppins";
+  font-size: 14px;
   cursor: pointer;
 }
 
 .MemberProfileLogout {
+  position: absolute;
+  top: 2%;
+  right: 1%;
   color: #2d285c;
   width: 35px;
   height: 35px;
